+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Not released yet
 
-* Allows repack without Castor logo e.g. `castor repack --show-logo=false`
+* Allows repack without Castor logo e.g. `castor repack --hide-logo`
 
 ## 0.22.0 (2024-12-30)